myproject/core/pagination.py
from collections import OrderedDict

from rest_framework.pagination import LimitOffsetPagination
from rest_framework.response import Response


class CustomPagination(LimitOffsetPagination):

    def get_paginated_response(self, data):
        return Response(OrderedDict([
            ('count', self.count),
            ('next', self.get_next_link()),
            ('offset', self.offset + self.limit),
            ('previous', self.get_previous_link()),
            ('results', data)
        ]))
