myproject/requirements.txt
asgiref==3.5.0
certifi==2021.10.8
charset-normalizer==2.0.12
coverage==6.3.2
Django==4.0.2
django-environ==0.8.1
djangorestframework==3.13.1
django-cors-headers==3.13.0
drf-nested-routers==0.93.4
environ==1.0
factory-boy==3.2.1
Faker==13.2.0
idna==3.3
# psycopg2==2.9.3
PyJWT==2.3.0
python-dateutil==2.8.2
pytz==2021.3
requests==2.27.1
shell==1.0.1
six==1.16.0
sqlparse==0.4.2
tqdm==4.62.3
urllib3==1.26.8
drf-yasg==1.20.0
django-storages==1.12.3
boto3==1.21
sentry-sdk==1.5.7
pep8==1.7.1
autopep8==1.6.0
gunicorn==20.1.0
dj-database-url==0.5.0
whitenoise==6.0.0
pre-commit==2.17.0
flake8==4.0.1